pyproject.toml
[tool.poetry]
name = "transcription-quality"
version = "0.1.0"
description = ""
authors = ["sofianeH1 <sofiane.hamiti18@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
scikit-learn = "^1.7.2"
jiwer = "^4.0.0"
python-dotenv = "^1.0.1"



[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
black = "^24.10.0"
ruff = "^0.6.9"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\\.(py)$'
extend-exclude = '(^/venv/|^/\\.venv/|^/__pycache__/|^/results/)'

[tool.isort]
profile = "black"
line_length = 100
skip = ["venv", ".venv", "__pycache__", "results"]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["venv", ".venv", "__pycache__", "results"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
